refactor(core): rename `withTable` member to avoid unwanted collition
table-wrapper component was implictly implementing the above mentioned
interface just by having the `pageChange$` and the `paginator`
properties while still wiring up the paginator and the page change
event.

This was causing the event to be fired too many times.

With this change we want to reduce to the chances of an unvolutary
implementation.